+++ scripts/benchmarking/parse_benchmark_logs.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python
+from __future__ import annotations
+
 import argparse
 import re
+import sys
+from collections import defaultdict
+from datetime import datetime
 from pathlib import Path
 
 import pandas as pd
@@ -26,28 +31,49 @@ def get_df(dolphin_config_file: Filename):
 
     log_file = Path(w.log_file)
 
-    result = _get_memory(log_file)
+    result = _parse_logfile(log_file)
     cfg_data = _parse_config(w)
 
     result.update(cfg_data)
     return pd.DataFrame([result])
 
 
-def _get_memory(logfile):
-    out = {"file": str(Path(logfile).resolve())}
+def _parse_logfile(logfile: Path) -> dict:
+    out: dict = defaultdict(list)
+    out["file"] = str(Path(logfile).resolve())
     mempat = r"Maximum memory usage: (\d\.\d{2}) GB"
     timepat = (
         r"Total elapsed time for dolphin.workflows.s1_disp.run : (\d*\.\d{2}) minutes"
         r" \((\d*\.\d{2}) seconds\)"
     )
-    for line in open(logfile).readlines():
-        m = re.search(mempat, line)
-        if m:
+    wrapped_phase_timepat = (
+        r"Total elapsed time for dolphin.workflows.wrapped_phase.run : (\d*\.\d{2})"
+        r" minutes"
+        r" \((\d*\.\d{2}) seconds\)"
+    )
+    cfg_version_pat = r"Config file dolphin version: (.*)"
+    run_version_pat = r"Current running dolphin version: (.*)"
+    # lines start with datetimes like 2023-04-19 17:13:02
+    # so at first line, parse and store the time of execution
+    for i, line in enumerate(open(logfile).readlines()):
+        if i == 0:
+            datetime_str = line[:19]
+            out["execution_time"] = datetime.strptime(datetime_str, "%Y-%m-%d %H:%M:%S")
+        if m := re.search(mempat, line):
             out["memory"] = float(m.groups()[0])
             continue
-        m = re.search(timepat, line)
-        if m:
+        if m := re.search(timepat, line):
             out["runtime"] = float(m.groups()[1])
+            continue
+        if m := re.search(wrapped_phase_timepat, line):
+            out["wrapped_phase_runtimes"].append(float(m.groups()[1]))
+            continue
+        if m := re.search(cfg_version_pat, line):
+            out["config_version"] = m.groups()[0]
+            continue
+        if m := re.search(run_version_pat, line):
+            out["run_version"] = m.groups()[0]
+            continue
     return out
 
 
@@ -60,7 +86,6 @@ def _parse_config(workflow: Workflow):
         "n_slc": len(workflow.cslc_file_list),
         "n_workers": workflow.worker_settings.n_workers,
         "creation_time": workflow.creation_time_utc,
-        "dolphin_version": workflow.dolphin_version,
     }
 
 
@@ -74,7 +99,7 @@ def _get_cli_args():
     parser.add_argument(
         "-o",
         "--outfile",
-        help="Output file (HTML). If None, outputs as `log_file`.html ",
+        help="Output file (CSV). If None, outputs as `log_file`.csv",
         default=None,
     )
     return parser.parse_args()
@@ -89,8 +114,6 @@ def main():
 
     df = pd.concat(dfs)
 
-    from datetime import datetime
-
     if not args.outfile:
         # Save as csv file with same directory as log_file
         outfile = (
@@ -100,8 +123,9 @@ def main():
     else:
         outfile = args.outfile
 
-    if not outfile.endswith(".html"):
-        raise ValueError("Output file must be csv")
+    if not outfile.endswith(".csv"):
+        outfile = outfile + ".csv"
+        print(f"Output file must be csv. Writing to {outfile}", file=sys.stderr)
 
     df.to_csv(outfile, index=False)